// src/app/layout/team/model/team.model.ts
export interface TeamModel {
    ID: number;
    image: string;
    fullname: string;
    title: string;
    collaboration: number;
    attitude_positive: number;
    customer_service: number;
    old_experience: number;
    email: string;
    phone: string;
    content: string;
    link_facebook: string;
    link_x: string;
    link_linkedin: string;
    link_youtube: string;
    signature: string;
}